move channelProcessor
